cambio posicion objetos al instanciar

+++ Assets/Scripts/GameManager.cs
@@ -8,7 +8,7 @@ public class GameManager : MonoBehaviour {
 
 	public static GameManager instance = null; //Estático para ser llamado por otro script
 	public int level = 1; //Nivel en el que empieza el juego
-	public float playerLife = 100;
+	public int playerLife = 100;
 	public MusicController musicSystem;
 
 	private BoardManager boardScript;
